backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# AI/ML
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0
sentence-transformers==2.3.1
chromadb==0.4.22
transformers==4.37.2
torch==2.1.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
beautifulsoup4==4.12.3
markdown==3.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
celery==5.3.6
